app/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
shapely
python-dotenv
xgboost